Fedger.io
This is a publicly available API comprising of several sub-APIs: company, fundings, geodata, news etc. 
The API itself is proprietary as the company behind charges for its usage but there's a certain amounts of "free calls" for every registered consumer.
Registering the API key is very easy and for free. I have also written a PureScript library to access the API (the backend logic is written in JS, so it can be easily used without PureScript as well).

Not sure if such APIs are welcome here, so feel free to reject this PR.